content/canvas/test/test_type.replace.html

<!DOCTYPE HTML>
<title>Canvas test: type.replace</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

window.HTMLCanvasElement.prototype.getContext = function (name) { return 0; };
ok(canvas.getContext('2d') === 0, "canvas.getContext('2d') === 0");

if (!_deferred) SimpleTest.finish();
});
</script>
